+++ CHANGELOG.md
@@ -15,6 +15,7 @@ file. This change log follows the conventions of
 - The `t310` and `t315` methods of `HubHandler` can now create `T31XHandler` handlers for either of the two device types.
 - The child device handlers for the H100 hub and the P300 power strip have been redesigned to eliminate the use of lifetimes, to facilitate FFI integrations.
 - The comments of `start_timestamp` and `end_timestamp` fields in `EnergyDataResult` have been updated to better describe their purpose.
+- `S200BRotationParams`'s field `degrees` has been renamed to `rotation_degrees`.
 
 ### Fixed
 
@@ -28,6 +29,7 @@ file. This change log follows the conventions of
 
 ### Added
 
+- Added full support for the S200B switches though the `S200BHandler` handler.
 - Added full support for the T100 sensors though the `T100Handler` handler.
 - Added full support for the T110 sensors though the `T110Handler` handler.
 - Added full support for the T300 sensors though the `T300Handler` handler.